PCHR-3027: (review) Rename variableService to siteInformation
