+++ uk.co.compucorp.civicrm.hrcore/CRM/HRCore/CMSData/SiteInformation/SiteInformationInterface.php
@@ -1,0 +1,15 @@
+<?php
+
+/**
+ * Implement this interface when providing a class to fetch CMS site info
+ */
+interface CRM_HRCore_CMSData_SiteInformation_SiteInformationInterface {
+
+  /**
+   * Gets the site name.
+   *
+   * @return string
+   */
+  public function getSiteName();
+
+}
